% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_graph.R
\name{calculate_type_counts}
\alias{calculate_type_counts}
\title{Calculate the counts before next division}
\usage{
calculate_type_counts(type_graph, init_count = type_graph$founder_size)
}
\description{
Calculate the counts before next division
}
