% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_n_identity}
\alias{check_n_identity}
\title{compare if two strings are identical other than N characters}
\usage{
check_n_identity(str_n, str_ref)
}
\description{
compare if two strings are identical other than N characters
}
