% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concat_mut_p}
\alias{concat_mut_p}
\title{Concatenate two `mut_p` (mutagenesis parameters) objects.}
\usage{
concat_mut_p(mut_p1, mut_p2)
}
\description{
Concatenate two `mut_p` (mutagenesis parameters) objects.
}
