% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concat_mut_p_list}
\alias{concat_mut_p_list}
\title{Concatenate a list of `mut_p` (mutagenesis parameters) objects.}
\usage{
concat_mut_p_list(mut_p_list)
}
\description{
Concatenate a list of `mut_p` (mutagenesis parameters) objects.
}
