% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{construct_true_lineage_multi}
\alias{construct_true_lineage_multi}
\title{deprecated now (repplaced by mod2)
deprecated but no mod2 available now}
\usage{
construct_true_lineage_multi(gens1, tip_id)
}
\description{
deprecated now (repplaced by mod2)
deprecated but no mod2 available now
}
