% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{distribute_all_mut_counts}
\alias{distribute_all_mut_counts}
\title{Distribute allele counts for all elements}
\usage{
distribute_all_mut_counts(mut_counts_list, n1, n2)
}
\description{
Distribute allele counts for all elements
}
