% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{distribute_mut_counts}
\alias{distribute_mut_counts}
\title{Distribute mutant allele counts between two groups}
\usage{
distribute_mut_counts(mut_counts, n1, n2)
}
\description{
Distribute mutant allele counts between two groups
}
