% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{edges_df_to_newick}
\alias{edges_df_to_newick}
\title{Convert simplified edges df to phylo}
\usage{
edges_df_to_newick(edges_df_simple, nodes_df)
}
\description{
Convert simplified edges df to phylo
}
