% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_graph.R
\name{generate_diff_outcome}
\alias{generate_diff_outcome}
\title{Generate differentiation outcome matrix for type graph}
\usage{
generate_diff_outcome(type_graph)
}
\description{
Generate differentiation outcome matrix for type graph
}
