% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_graph.R
\name{generate_edge_df}
\alias{generate_edge_df}
\title{Generate edge data.frame for type graph, getting edge lengths from laid out graph}
\usage{
generate_edge_df(type_graph)
}
\description{
Generate edge data.frame for type graph, getting edge lengths from laid out graph
}
