% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{generate_mut_ids}
\alias{generate_mut_ids}
\title{Generate a batch of depth-1 mutante allele}
\usage{
generate_mut_ids(n, mut_element, recur_ver = TRUE)
}
\description{
Generate a batch of depth-1 mutante allele
}
