% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{generate_nonrecur_mut_ids}
\alias{generate_nonrecur_mut_ids}
\title{Generate a batch of non recurring depth-1 mutante allele}
\usage{
generate_nonrecur_mut_ids(n, mut_element)
}
\description{
Generate a batch of non recurring depth-1 mutante allele
}
