% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{generate_recur_mut_ids}
\alias{generate_recur_mut_ids}
\title{Generate a batch of recurring depth-1 mutant allele}
\usage{
generate_recur_mut_ids(n, mut_element, version = TRUE)
}
\description{
Generate a batch of recurring depth-1 mutant allele
}
