% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny_model_v2.R
\name{get_cell_type_counts}
\alias{get_cell_type_counts}
\title{current version of cell phylogeny model}
\usage{
get_cell_type_counts(type_graph, tp)
}
\description{
current version of cell phylogeny model
}
