% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_fase_multifurc.R
\name{get_edge_diff}
\alias{get_edge_diff}
\title{ICE-FASE inference when topology has multifurcations}
\usage{
get_edge_diff(x, data_obj, trans_time)
}
\description{
ICE-FASE inference when topology has multifurcations
}
