% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_func.R
\name{get_edge_len}
\alias{get_edge_len}
\title{get edge length statistics to distinguish between g0 and g1}
\usage{
get_edge_len(phy_list)
}
\description{
get edge length statistics to distinguish between g0 and g1
}
