% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{get_mut_duration}
\alias{get_mut_duration}
\title{get overlaps of intervals using trick on IRanges
TODO: dependes on IRanges}
\usage{
get_mut_duration(active_time, start_time, end_time)
}
\description{
get overlaps of intervals using trick on IRanges
TODO: dependes on IRanges
}
