% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{get_next}
\alias{get_next}
\title{extend out node to the next bifurcation in the lineage tree}
\usage{
get_next(y, edges_list)
}
\description{
extend out node to the next bifurcation in the lineage tree
}
