% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaulate_experiment.R
\name{get_node_mapping}
\alias{get_node_mapping}
\title{get mapping of inferred progenitor states to true progenitor states}
\usage{
get_node_mapping(data_obj, type_graph)
}
\description{
get mapping of inferred progenitor states to true progenitor states
}
