% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{get_t_total}
\alias{get_t_total}
\title{deprecated function, used for somatic mutation
get the total amount of time passed that the site could have mutated}
\usage{
get_t_total(active_time)
}
\description{
deprecated function, used for somatic mutation
get the total amount of time passed that the site could have mutated
}
