% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny_model_v2.R
\name{get_true_size_mod2}
\alias{get_true_size_mod2}
\title{this function returns the incoming size as well as the outgoing sizes}
\usage{
get_true_size_mod2(type_graph, delay = F)
}
\description{
this function returns the incoming size as well as the outgoing sizes
}
