% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_fase.R
\name{impute_characters}
\alias{impute_characters}
\title{Impute missing characters from cell allele matrix}
\usage{
impute_characters(mat, nrounds = 100, max_depth = 4)
}
\arguments{
\item{mat}{character matrix with rows corresponding to cells and columns corresponding to barcoding sites}
}
\description{
Impute missing characters from cell allele matrix
}
