% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{init_mut_counts_list}
\alias{init_mut_counts_list}
\title{Initialize mutant allele count list for bulk counting}
\usage{
init_mut_counts_list(mut_param, num_cells)
}
\description{
Initialize mutant allele count list for bulk counting
}
