% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{initialize_mut_element_counter}
\alias{initialize_mut_element_counter}
\title{Initialize the global mut_element counter for non-recurring alleles}
\usage{
initialize_mut_element_counter(mut_element, element_dim = 1)
}
\description{
Initialize the global mut_element counter for non-recurring alleles
}
