% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{initialize_mut_state}
\alias{initialize_mut_state}
\title{Initialize unmutated matrix for single cell model}
\usage{
initialize_mut_state(mut_param, num_cells)
}
\description{
Initialize unmutated matrix for single cell model
}
