% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_zero_range}
\alias{is_zero_range}
\title{Check if a vector is all the same value}
\usage{
is_zero_range(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{An input vector.}

\item{tol}{tolerance when comparing elements of the vector}
}
\description{
Check if a vector is all the same value
}
