% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_fase.R
\name{make_gr_tr_data}
\alias{make_gr_tr_data}
\title{mostly deprecated functions of ICE_FASE, replaced by 'ice_fase_mod1.R'
some functions are still used for processing downstream results}
\usage{
make_gr_tr_data(gr, tr, sc_celltypes)
}
\description{
mostly deprecated functions of ICE_FASE, replaced by 'ice_fase_mod1.R'
some functions are still used for processing downstream results
}
