% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_heatmap_col}
\alias{make_heatmap_col}
\title{Make a color palette for Heatmap visualization}
\usage{
make_heatmap_col(
  max_val,
  mid_val = max_val/2,
  min_val = 0,
  col = c("blue", "white", "red")
)
}
\description{
Make a color palette for Heatmap visualization
}
