% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{make_mut_element}
\alias{make_mut_element}
\title{Constructor for mutation parameter for a single element
Only supports max_depth of 1}
\usage{
make_mut_element(
  element_id,
  mut_rate,
  active_time,
  recur_prob = 0,
  recur_vec = NULL,
  max_depth = 1
)
}
\arguments{
\item{recur}{chance of recurring}
}
\description{
Constructor for mutation parameter for a single element
Only supports max_depth of 1
}
