% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{make_mut_param_by_rate}
\alias{make_mut_param_by_rate}
\title{Make a mut_param by a sequence of mutation rates}
\usage{
make_mut_param_by_rate(mut_rate, active_time, recur_prob = 0, recur_vec = NULL)
}
\description{
Make a mut_param by a sequence of mutation rates
}
