% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{make_mut_param_by_rate_rvec}
\alias{make_mut_param_by_rate_rvec}
\title{Make a mut_param by a sequence of mutation rates and multinomial allele probabilities}
\usage{
make_mut_param_by_rate_rvec(
  mut_rate,
  active_time,
  recur_prob_vec,
  recur_vec_list
)
}
\description{
Make a mut_param by a sequence of mutation rates and multinomial allele probabilities
}
