% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_graph.R
\name{make_type_graph}
\alias{make_type_graph}
\title{Constructor for type graph}
\usage{
make_type_graph(
  root_id,
  node_id,
  tip_id,
  merge_matrix,
  differntiation_time,
  differntiation_mode_probs,
  double_time,
  founder_size,
  target_time,
  transition_selection = NULL
)
}
\arguments{
\item{root_id}{name of root, which is the integer num_tip - 1}

\item{node_id}{vector of names of the graph nodes, i.e. progenitor states}

\item{tip_id}{vector of names of the graph tip, i.e. terminal states}

\item{merge_matrix}{character matrix specifying the sequence of merges}

\item{differntiation_mode_probs}{commitment bias, need to be vector of length 5, only first values two are non-zero}

\item{double_time}{list of doubling times for each state}

\item{founder_size}{founder size}

\item{target_time}{time of sample collection}

\item{differentiation_time}{commitment time of states}
}
\description{
Constructor for type graph
}
