% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutate_barcodes}
\alias{mutate_barcodes}
\title{Generate mutant allele for single cell mutation states}
\usage{
mutate_barcodes(mut_state, life_duration, mut_param)
}
\description{
Generate mutant allele for single cell mutation states
}
