% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{normalize_tree}
\alias{normalize_tree}
\title{Construct lineage tree from sim_history}
\usage{
normalize_tree(tree, check.ultrametric = TRUE)
}
\description{
Construct lineage tree from sim_history
}
