% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny_model.R
\name{num_gen_single_type}
\alias{num_gen_single_type}
\title{Given a single cell type, generate the number of generations it will divide until it reaches target time or differentiate}
\usage{
num_gen_single_type(cell_type, rtime, type_graph)
}
\description{
Given a single cell type, generate the number of generations it will divide until it reaches target time or differentiate
}
