% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotime.R
\name{phylotime}
\alias{phylotime}
\title{Phylotime for reconstructing time-scaled phylogeny based on lineage barcodes}
\usage{
phylotime(sc_mat, t_total, mut_p = NULL, return_dist = F)
}
\arguments{
\item{sc_mat}{character matrix cell x barcoding sites}

\item{t_total}{total amount of time since barcode activation}

\item{return_dist}{whetehr to return pairwise distance between cells, by default, an ape tree is returned}
}
\description{
Phylotime for reconstructing time-scaled phylogeny based on lineage barcodes
}
