% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_fase.R
\name{plot_gr_data}
\alias{plot_gr_data}
\title{plot ice_fase results}
\usage{
plot_gr_data(out_data, target_time, gr_col = NULL, gr_lab = NULL)
}
\arguments{
\item{out_data}{"ice_fase_results" object}
}
\description{
plot ice_fase results
}
