% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_graph.R
\name{plot_type_graph}
\alias{plot_type_graph}
\title{Plot type graph using ggraph}
\usage{
plot_type_graph(
  type_graph,
  node_col_mapper = white_col_mapper,
  node_label_mapper = NULL,
  effective_time = T,
  show_node_text = T
)
}
\description{
Plot type graph using ggraph
}
