% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{push_pdf}
\alias{push_pdf}
\title{Push a ggplot object to a PDF file}
\usage{
push_pdf(
  g,
  file_name,
  width = 4,
  height = 4,
  ps = 12,
  dir = "./",
  open_file = T
)
}
\description{
Push a ggplot object to a PDF file
}
