% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{push_png}
\alias{push_png}
\title{Push a ggplot object to a PNG file}
\usage{
push_png(
  g,
  file_name,
  width = 4,
  height = 4,
  ps = 12,
  res = 300,
  dir = "../LTModelPlots/",
  open_file = T
)
}
\description{
Push a ggplot object to a PNG file
}
