% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_frac}
\alias{round_frac}
\title{Assign counts based on total and fraction, rounded to integer}
\usage{
round_frac(total, prob)
}
\description{
Assign counts based on total and fraction, rounded to integer
}
