% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rstoc}
\alias{rstoc}
\title{Round a number stochasticly, returns integer greater than or equal to one.}
\usage{
rstoc(x)
}
\arguments{
\item{x}{integer}
}
\description{
Round a number stochasticly, returns integer greater than or equal to one.
}
