% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{sample_all_mut_counts}
\alias{sample_all_mut_counts}
\title{Draw one sample of mutation counts for all elements}
\usage{
sample_all_mut_counts(
  mut_counts_list,
  mut_param,
  start_time,
  end_time,
  double_counts = F
)
}
\description{
Draw one sample of mutation counts for all elements
}
