% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{sample_all_mut_counts_gen}
\alias{sample_all_mut_counts_gen}
\title{Draw one sample of the mutation counts after a number of generatioins}
\usage{
sample_all_mut_counts_gen(
  mut_counts_list,
  mut_param,
  life_duration,
  num_gen,
  double_counts = T
)
}
\arguments{
\item{num_gen}{the number of generation}
}
\description{
Draw one sample of the mutation counts after a number of generatioins
}
