% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_sample.R
\name{sample_doublet_branch}
\alias{sample_doublet_branch}
\title{the stochastic coal function
sample the number of doublet branches when dividing n cells into 2*n cell}
\usage{
sample_doublet_branch(n, k)
}
\description{
the stochastic coal function
sample the number of doublet branches when dividing n cells into 2*n cell
}
