% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{sample_mut_counts}
\alias{sample_mut_counts}
\title{Draw one sample of the mutation counts after some time}
\usage{
sample_mut_counts(mut_counts, mut_element, duration)
}
\arguments{
\item{life_duration}{length of the generation}

\item{double_counts}{whether the counts of the alleles should be doubled as a reulst of division}
}
\description{
Draw one sample of the mutation counts after some time
}
