% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{sim_history_to_edges}
\alias{sim_history_to_edges}
\title{Convert sim_history to edges data.frame}
\usage{
sim_history_to_edges(sim_history)
}
\description{
Convert sim_history to edges data.frame
}
