% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{sim_history_to_nodes}
\alias{sim_history_to_nodes}
\title{Convert sim_history to nodes data.frame}
\usage{
sim_history_to_nodes(sim_history)
}
\description{
Convert sim_history to nodes data.frame
}
