% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_cell.R
\name{simplify_edges_df}
\alias{simplify_edges_df}
\title{Remove internal nodes with single daughter}
\usage{
simplify_edges_df(edges_df, nodes_df)
}
\description{
Remove internal nodes with single daughter
}
