% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{subsample_all_mut_counts}
\alias{subsample_all_mut_counts}
\title{Draw subsample of allele counts for all elements}
\usage{
subsample_all_mut_counts(mut_counts_list, sample_size)
}
\description{
Draw subsample of allele counts for all elements
}
