% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{subsample_mut_counts}
\alias{subsample_mut_counts}
\title{Draw a subsample of mutant allele counts}
\usage{
subsample_mut_counts(mut_counts, sample_size)
}
\description{
Draw a subsample of mutant allele counts
}
