% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_counts}
\alias{scale_counts}
\title{Scale counts based on a factor, rounded to integer, counts must be at least 1}
\usage{
scale_counts(counts, z)
}
\description{
Scale counts based on a factor, rounded to integer, counts must be at least 1
}
